+++ tests/test_base_device.py
@@ -20,7 +20,6 @@ import numpy as np
 import pytest
 from semantic_version import Version
 import qiskit_ibm_runtime
-import qiskit
 
 import pennylane as qml
 from pennylane.tape.qscript import QuantumScript
@@ -142,26 +141,22 @@ class TestSupportForV1andV2:
     )
     def test_v1_and_v2_mocked(self, backend):
         """Test that device initializes with no error mocked"""
-        dev = QiskitDevice2(wires=10, backend=backend, use_primitives=True)
+        dev = QiskitDevice2(wires=10, backend=backend)
         assert dev._backend == backend
 
     @pytest.mark.parametrize(
-        "backend, use_primitives, shape",
+        "backend, shape",
         [
-            (FakeManila(), True, (1, 1024)),
-            (FakeManila(), False, (1024,)),
-            (FakeManilaV2(), True, (1, 1024)),
-            (FakeManilaV2(), False, (1024,)),
+            (FakeManila(), (1, 1024)),
+            (FakeManilaV2(), (1, 1024)),
         ],
     )
-    @pytest.mark.skipif(
-        Version(qiskit.__version__) < Version("1.0.0"),
-        reason="Session initialization is not supported for local simulators for Qiskit version < 1.0/qiskit_ibm_runtime version < 0.22.0",
-        ## See https://docs.quantum.ibm.com/api/migration-guides/local-simulators for additional details
+    @pytest.mark.skip(
+        reason="The functionality of using sampler to get an accurate answer is not yet implemented"
     )
-    def test_v1_and_v2_manila(self, backend, use_primitives, shape):
+    def test_v1_and_v2_manila(self, backend, shape):
         """Test that device initializes and runs without error with V1 and V2 backends by Qiskit"""
-        dev = QiskitDevice2(wires=5, backend=backend, use_primitives=use_primitives)
+        dev = QiskitDevice2(wires=5, backend=backend)
 
         @qml.qnode(dev)
         def circuit(x):
@@ -190,12 +185,6 @@ class TestDeviceInitialization:
 
         assert dev2._compile_backend != dev2._backend
         assert dev2._compile_backend == compile_backend
-
-    @pytest.mark.parametrize("use_primitives", [True, False])
-    def test_use_primitives_kwarg(self, use_primitives):
-        """Test the _use_primitives attribute is set on initialization"""
-        dev = QiskitDevice2(wires=2, backend=backend, use_primitives=use_primitives)
-        assert dev._use_primitives == use_primitives
 
     def test_no_shots_warns_and_defaults(self):
         """Test that initializing with shots=None raises a warning indicating that
@@ -325,8 +314,7 @@ class TestDevicePreprocessing:
                 ],
                 [
                     [qml.expval(qml.PauliZ(1)), qml.var(qml.PauliY(0))],
-                    [qml.probs(wires=[2])],
-                    [qml.counts()],
+                    [qml.counts(), qml.probs(wires=[2])],
                 ],
             ),
             (
@@ -390,7 +378,7 @@ class TestDevicePreprocessing:
     @pytest.mark.filterwarnings("ignore::UserWarning")
     def test_split_execution_types(self, measurements, expectation):
         """Test that the split_execution_types transform splits measurements into Estimator-based
-        (expval, var), Sampler-based (probs) and raw-sample based (everything else)"""
+        (expval, var) and Sampler-based (everything else)"""
 
         operations = [qml.PauliX(0), qml.PauliY(1), qml.Hadamard(2), qml.CNOT([2, 1])]
         qs = QuantumScript(operations, measurements=measurements)
@@ -439,17 +427,17 @@ class TestDevicePreprocessing:
         [
             ([qml.expval(qml.PauliZ(0)), qml.probs(wires=[0, 1])], 2),
             ([qml.expval(qml.PauliZ(0)), qml.sample(wires=[0, 1])], 2),
-            ([qml.counts(), qml.probs(wires=[0, 1]), qml.sample()], 2),
+            ([qml.counts(), qml.probs(wires=[0, 1]), qml.sample()], 1),
             ([qml.var(qml.PauliZ(0)), qml.expval(qml.PauliX(1))], 1),
-            ([qml.probs(wires=[0]), qml.counts(), qml.var(qml.PauliY(2))], 3),
+            ([qml.probs(wires=[0]), qml.counts(), qml.var(qml.PauliY(2))], 2),
         ],
     )
     def test_preprocess_splits_incompatible_primitive_measurements(self, measurements, num_types):
         """Test that the default behaviour for preprocess it to split the tapes based
-        on meausrement type. Expval and Variance are one type (Estimator), Probs another (Sampler),
-        and everything else a third (raw sample-based measurements)."""
+        on meausrement type. Expval and Variance are one type (Estimator), Probs and raw-sample based measurements
+        are another type (Sampler)."""
 
-        dev = QiskitDevice2(wires=5, backend=backend, use_primitives=True)
+        dev = QiskitDevice2(wires=5, backend=backend)
         qs = QuantumScript([], measurements=measurements, shots=qml.measurements.Shots(1000))
 
         program, _ = dev.preprocess()
@@ -457,29 +445,6 @@ class TestDevicePreprocessing:
 
         # measurements that are incompatible are split when use_primtives=True
         assert len(tapes) == num_types
-
-    @pytest.mark.parametrize(
-        "measurements",
-        [
-            [qml.expval(qml.PauliZ(0)), qml.probs(wires=[0, 1])],
-            [qml.expval(qml.PauliZ(0)), qml.sample(wires=[0, 1])],
-            [qml.counts(), qml.probs(wires=[0, 1]), qml.sample()],
-        ],
-    )
-    def test_preprocess_measurements_without_primitives(self, measurements):
-        """Test if Primitives are not being used that the preprocess does not split
-        the tapes based on measurement type"""
-
-        qs = QuantumScript([], measurements=measurements, shots=qml.measurements.Shots(1000))
-
-        dev = QiskitDevice2(wires=5, backend=backend, use_primitives=False)
-        program, _ = dev.preprocess()
-
-        tapes, _ = program([qs])
-
-        # measurements that are incompatible on the primitive-based device
-        # are not split when use_primtives=False
-        assert len(tapes) == 1
 
     def test_preprocess_decomposes_unsupported_operator(self):
         """Test that the device preprocess decomposes operators that
@@ -833,47 +798,12 @@ class TestMockedExecution:
         assert len(np.argwhere([np.allclose(s, [0, 1]) for s in samples])) == results_dict["10"]
         assert len(np.argwhere([np.allclose(s, [1, 0]) for s in samples])) == results_dict["01"]
 
-    @pytest.mark.parametrize("backend", [backend, legacy_backend])
-    def test_execute_pipeline_no_primitives_mocked(self, mocker, backend):
-        """Test that a device **not** using Primitives only calls the _execute_runtime_service
-        to execute, regardless of measurement type"""
-
-        dev = QiskitDevice2(
-            wires=5, backend=backend, use_primitives=False, session=MockSession(backend)
-        )
-
-        initial_session = dev._session
-
-        sampler_execute = mocker.spy(dev, "_execute_sampler")
-        estimator_execute = mocker.spy(dev, "_execute_estimator")
-
-        qs = QuantumScript(
-            [qml.PauliX(0), qml.PauliY(1)],
-            measurements=[
-                qml.expval(qml.PauliZ(0)),
-                qml.probs(wires=[0, 1]),
-                qml.counts(),
-                qml.sample(),
-            ],
-        )
-
-        with patch.object(dev, "_execute_runtime_service", return_value="runtime_execute_res"):
-            runtime_service_execute = mocker.spy(dev, "_execute_runtime_service")
-            res = dev.execute(qs)
-
-        runtime_service_execute.assert_called_once()
-        sampler_execute.assert_not_called()
-        estimator_execute.assert_not_called()
-
-        assert res == "runtime_execute_res"
-        assert initial_session == dev._session  # session is not changed
-
     @patch("pennylane_qiskit.qiskit_device2.QiskitDevice2._execute_estimator")
     def test_execute_pipeline_primitives_no_session(self, mocker):
         """Test that a Primitives-based device initialized with no Session creates one for the
         execution, and then returns the device session to None."""
 
-        dev = QiskitDevice2(wires=5, backend=backend, use_primitives=True, session=None)
+        dev = QiskitDevice2(wires=5, backend=backend, session=None)
 
         assert dev._session is None
 
@@ -887,13 +817,10 @@ class TestMockedExecution:
 
     @pytest.mark.parametrize("backend", [backend, legacy_backend])
     def test_execute_pipeline_with_all_execute_types_mocked(self, mocker, backend):
-        """Test that a device that **is** using Primitives calls the _execute_runtime_service
-        to execute measurements that require raw samples, and the relevant primitive measurements
-        on the other measurements"""
+        """Test that a device executes measurements that require raw samples on the sampler,
+        and the relevant primitive measurements on the estimator"""
 
-        dev = QiskitDevice2(
-            wires=5, backend=backend, use_primitives=True, session=MockSession(backend)
-        )
+        dev = QiskitDevice2(wires=5, backend=backend, session=MockSession(backend))
 
         qs = QuantumScript(
             [qml.PauliX(0), qml.PauliY(1)],
@@ -906,23 +833,19 @@ class TestMockedExecution:
         )
         tapes, _ = split_execution_types(qs)
 
-        with patch.object(dev, "_execute_runtime_service", return_value="runtime_execute_res"):
-            with patch.object(dev, "_execute_sampler", return_value="sampler_execute_res"):
-                with patch.object(dev, "_execute_estimator", return_value="estimator_execute_res"):
-                    runtime_service_execute = mocker.spy(dev, "_execute_runtime_service")
-                    sampler_execute = mocker.spy(dev, "_execute_sampler")
-                    estimator_execute = mocker.spy(dev, "_execute_estimator")
+        with patch.object(dev, "_execute_sampler", return_value="sampler_execute_res"):
+            with patch.object(dev, "_execute_estimator", return_value="estimator_execute_res"):
+                sampler_execute = mocker.spy(dev, "_execute_sampler")
+                estimator_execute = mocker.spy(dev, "_execute_estimator")
 
-                    res = dev.execute(tapes)
+                res = dev.execute(tapes)
 
-        runtime_service_execute.assert_called_once()
         sampler_execute.assert_called_once()
         estimator_execute.assert_called_once()
 
         assert res == [
             "estimator_execute_res",
             "sampler_execute_res",
-            "runtime_execute_res",
         ]
 
     @patch("pennylane_qiskit.qiskit_device2.Estimator")
@@ -954,45 +877,10 @@ class TestMockedExecution:
         # to emphasize, this did nothing except appease CodeCov
         assert isinstance(result[0], Mock)
 
-    @patch("pennylane_qiskit.qiskit_device2.transpile")
-    def test_execute_runtime_service_mocked(self, mocked_transpile):
-        """Test the _execute_sampler function using a mocked version of Sampler
-        that returns a meaningless result."""
-
-        dev = QiskitDevice2(wires=5, backend=backend, use_primitives=True)
-
-        mock_counts = {"00": 125, "10": 500, "01": 250, "11": 125}
-        mock_result = Mock()
-        mock_job = Mock()
-        mock_service = Mock()
-
-        mock_result.get_counts = Mock(return_value=mock_counts)
-        mock_job.result = Mock(return_value=mock_result)
-        mock_service.run = Mock(return_value=mock_job)
-
-        dev._service = mock_service
-
-        qs = QuantumScript([qml.PauliX(0)], measurements=[qml.sample()])
-        result = dev._execute_runtime_service(qs, MockSession(backend))
-
-        samples = result[0]
-
-        assert len(samples) == sum(mock_counts.values())
-        assert len(samples[0]) == 2
-
-        assert len(np.argwhere([np.allclose(s, [0, 0]) for s in samples])) == mock_counts["00"]
-        assert len(np.argwhere([np.allclose(s, [1, 1]) for s in samples])) == mock_counts["11"]
-
-        # order of samples is swapped compared to keys (Qiskit wire order convention is reverse of PennyLane)
-        assert len(np.argwhere([np.allclose(s, [0, 1]) for s in samples])) == mock_counts["10"]
-        assert len(np.argwhere([np.allclose(s, [1, 0]) for s in samples])) == mock_counts["01"]
-
     def test_shot_vector_warning_mocked(self):
         """Test that a device that executes a circuit with an array of shots raises the appropriate warning"""
 
-        dev = QiskitDevice2(
-            wires=5, backend=backend, use_primitives=True, session=MockSession(backend)
-        )
+        dev = QiskitDevice2(wires=5, backend=backend, session=MockSession(backend))
         qs = QuantumScript(
             measurements=[
                 qml.expval(qml.PauliX(0)),
@@ -1008,11 +896,6 @@ class TestMockedExecution:
                 dev.execute(qs)
 
 
-@pytest.mark.skipif(
-    Version(qiskit.__version__) < Version("1.0.0"),
-    reason="Session initialization is not supported for local simulators for Qiskit version < 1.0/qiskit_ibm_runtime version < 0.22.0",
-    ## See https://docs.quantum.ibm.com/api/migration-guides/local-simulators for additional details
-)
 class TestExecution:
 
     @pytest.mark.parametrize("wire", [0, 1])
@@ -1033,9 +916,8 @@ class TestExecution:
         correspond correctly (wire ordering convention in Qiskit and PennyLane don't match.)
         """
 
-        dev = QiskitDevice2(wires=5, backend=backend, use_primitives=True)
+        dev = QiskitDevice2(wires=5, backend=backend)
 
-        runtime_service_execute = mocker.spy(dev, "_execute_runtime_service")
         sampler_execute = mocker.spy(dev, "_execute_sampler")
         estimator_execute = mocker.spy(dev, "_execute_estimator")
 
@@ -1053,7 +935,6 @@ class TestExecution:
 
         res = dev.execute(qs)
 
-        runtime_service_execute.assert_not_called()
         sampler_execute.assert_not_called()
         estimator_execute.assert_called_once()
 
@@ -1093,9 +974,8 @@ class TestExecution:
         """
 
         pl_dev = qml.device("default.qubit", wires=[0, 1, 2, 3])
-        dev = QiskitDevice2(wires=[0, 1, 2, 3], backend=backend, use_primitives=True)
+        dev = QiskitDevice2(wires=[0, 1, 2, 3], backend=backend)
 
-        runtime_service_execute = mocker.spy(dev, "_execute_runtime_service")
         sampler_execute = mocker.spy(dev, "_execute_sampler")
         estimator_execute = mocker.spy(dev, "_execute_estimator")
 
@@ -1111,7 +991,6 @@ class TestExecution:
         res = dev.execute(qs)
         expectation = pl_dev.execute(qs)
 
-        runtime_service_execute.assert_not_called()
         sampler_execute.assert_not_called()
         estimator_execute.assert_called_once()
 
@@ -1176,6 +1055,7 @@ class TestExecution:
 
     @pytest.mark.parametrize("num_wires", [1, 3, 5])
     @pytest.mark.parametrize("num_shots", [50, 100])
+    @pytest.mark.skip(reason="Need to replace this with using SamplerV2.")
     def test_generate_samples(self, num_wires, num_shots):
         qs = QuantumScript([], measurements=[qml.expval(qml.PauliX(0))])
 
@@ -1204,29 +1084,9 @@ class TestExecution:
         # nothing else is in samples
         assert [s for s in samples if not s in np.array([exp_res0, exp_res1])] == []
 
-    def test_tape_shots_used_runtime_service(self, mocker):
-        """Tests that device uses tape shots rather than device shots for _execute_runtime_service"""
-        dev = QiskitDevice2(wires=5, backend=backend, shots=2, use_primitives=True)
-
-        runtime_service_execute = mocker.spy(dev, "_execute_runtime_service")
-
-        @qml.qnode(dev)
-        def circuit():
-            return qml.sample()
-
-        res = circuit(shots=[5])
-
-        runtime_service_execute.assert_called_once()
-
-        assert len(res[0]) == 5
-
-        # Should reset to device shots if circuit ran again without shots defined
-        res = circuit()
-        assert len(res[0]) == 2
-
     def test_tape_shots_used_for_estimator(self, mocker):
         """Tests that device uses tape shots rather than device shots for estimator"""
-        dev = QiskitDevice2(wires=5, backend=backend, shots=2, use_primitives=True)
+        dev = QiskitDevice2(wires=5, backend=backend, shots=2)
 
         estimator_execute = mocker.spy(dev, "_execute_estimator")
 
@@ -1245,7 +1105,7 @@ class TestExecution:
 
     def test_tape_shots_used_for_sampler(self, mocker):
         """Tests that device uses tape shots rather than device shots for sampler"""
-        dev = QiskitDevice2(wires=5, backend=backend, shots=2, use_primitives=True)
+        dev = QiskitDevice2(wires=5, backend=backend, shots=2)
 
         sampler_execute = mocker.spy(dev, "_execute_sampler")
 
@@ -1265,7 +1125,7 @@ class TestExecution:
 
     def test_warning_for_shot_vector(self):
         """Tests that a warning is raised if a shot vector is passed and total shots of tape is used instead."""
-        dev = QiskitDevice2(wires=5, backend=backend, shots=2, use_primitives=True)
+        dev = QiskitDevice2(wires=5, backend=backend, shots=2)
 
         @qml.qnode(dev)
         def circuit():
@@ -1291,16 +1151,18 @@ class TestExecution:
         ],
     )
     @pytest.mark.filterwarnings("ignore::UserWarning")
+    @pytest.mark.skip(
+        reason="The functionality of using sampler to get the accurate answer is not yet implemented"
+    )
     def test_no_pauli_observable_gives_accurate_answer(self, mocker, observable):
-        """Test that the device uses _execute_runtime_service and _execute_estimator appropriately
+        """Test that the device uses _sampler and _execute_estimator appropriately
         and provides an accurate answer for measurements with observables that don't have a pauli_rep.
         """
 
-        dev = QiskitDevice2(wires=5, backend=backend, use_primitives=True)
+        dev = QiskitDevice2(wires=5, backend=backend)
 
         pl_dev = qml.device("default.qubit", wires=5)
 
-        runtime_service_execute = mocker.spy(dev, "_execute_runtime_service")
         estimator_execute = mocker.spy(dev, "_execute_estimator")
         sampler_execute = mocker.spy(dev, "_execute_sampler")
 
@@ -1319,9 +1181,8 @@ class TestExecution:
         res = circuit()
         pl_res = pl_circuit()
 
-        runtime_service_execute.assert_called_once()
         estimator_execute.assert_called_once()
-        sampler_execute.assert_not_called()
+        sampler_execute.assert_called_once()
 
         assert np.allclose(res, pl_res, atol=0.1)
 
@@ -1329,11 +1190,7 @@ class TestExecution:
         """Test that a warning is raised when device is passed a measurement on
         an observable that does not have a pauli_rep."""
 
-        dev = QiskitDevice2(
-            wires=5,
-            backend=backend,
-            use_primitives=True,
-        )
+        dev = QiskitDevice2(wires=5, backend=backend)
 
         @qml.qnode(dev)
         def circuit():